requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
spotipy==2.23.0
python-dotenv==1.0.1
jinja2==3.1.4
itsdangerous==2.2.0
starlette==0.37.2
httpx==0.27.0
aiofiles==23.2.1
